pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "isaac"
version = "0.1.0"
description = "Intelligent System for Autonomous Action and Cognition — neuro-symbolic autonomous agent"
readme = "README.md"
license = "AGPL-3.0-or-later"
requires-python = ">=3.10"
authors = [{ name = "Matheus Soran" }]
keywords = ["agent", "agi", "neuro-symbolic", "langgraph", "code-agent"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "langgraph>=0.2",
    "langchain-core>=0.3",
    "langchain-openai>=0.2",
    "langchain-anthropic>=0.2",
    "langchain-ollama>=0.2",
    "docker>=7.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "numpy>=1.26",
    "chromadb>=0.5",
    "networkx>=3.2",
    "typer>=0.12",
    "duckduckgo-search>=6.0",
    "python-telegram-bot>=21.0",
    "apscheduler>=3.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
]
browser = [
    "playwright>=1.44",
]
calendar = [
    "caldav>=1.3",
]

[project.scripts]
isaac = "isaac.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/isaac"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
